+++ src/components/layout/LayoutHeader.tsx
@@ -3,9 +3,14 @@ import React from 'react';
 import { ConfigProvider, Flex, Layout, Menu } from 'antd';
 import { useTranslation } from 'react-i18next';
 import Link from 'next/link';
+import styled from '@emotion/styled';
 import { COLORS } from 'src/styles/variables';
 import { Logo } from '~components/ui/logo';
 import { LangSwitch } from '~components/widgets/lang-switch';
+
+const MenuLabel = styled(Link)`
+    font-size: 16px;
+`;
 
 export const LayoutHeader = () => {
     const { t } = useTranslation();
@@ -13,25 +18,25 @@ export const LayoutHeader = () => {
     const menuItems = [
         {
             label: (
-                <Link href="/home" passHref>
+                <MenuLabel href="/home" passHref>
                     {t('header_menu_home')}
-                </Link>
+                </MenuLabel>
             ),
             key: 1,
         },
         {
             label: (
-                <Link href="/application" passHref>
+                <MenuLabel href="/application" passHref>
                     {t('header_menu_application')}
-                </Link>
+                </MenuLabel>
             ),
             key: 2,
         },
         {
             label: (
-                <Link href="/contacts" passHref>
+                <MenuLabel href="/contacts" passHref>
                     {t('header_menu_contacts')}
-                </Link>
+                </MenuLabel>
             ),
             key: 3,
         },
@@ -48,8 +53,8 @@ export const LayoutHeader = () => {
                         itemBg: 'none',
                         itemColor: `${COLORS.white}`,
                         itemHoverColor: `${COLORS.white}`,
-                        horizontalItemHoverBg: `${COLORS.pink}`,
-                        horizontalItemSelectedBg: `${COLORS.pink}`,
+                        horizontalItemHoverBg: `${COLORS.blueHovered}`,
+                        horizontalItemSelectedBg: `${COLORS.blueHovered}`,
                         horizontalItemSelectedColor: `${COLORS.white}`,
                         activeBarHeight: 0,
                         popupBg: COLORS.blue,
@@ -83,15 +88,17 @@ export const LayoutHeader = () => {
                         }}
                         items={menuItems}
                     />
-                    <LangSwitch />
-                    <Logo
-                        link="/"
-                        src={'/cat.jpg'}
-                        alt="music-studio"
-                        width="50"
-                        height="50"
-                        style={{ marginRight: '5px' }}
-                    />
+                    <Flex gap={20} align="center">
+                        <LangSwitch />
+                        <Logo
+                            link="/"
+                            src={'/logo.png'}
+                            alt="music-studio"
+                            width="50"
+                            height="50"
+                            style={{ marginRight: '5px' }}
+                        />
+                    </Flex>
                 </Flex>
             </Layout.Header>
         </ConfigProvider>
