--- src/components/ui/hook-form/button/index.tsx
@@ -1,12 +1,0 @@
-import { Button as ButtonAntd } from 'antd';
-import styled from '@emotion/styled';
-
-export const Button = styled(ButtonAntd)`
-    min-width: 150px;
-    min-height: 50px;
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    font-size: 16px;
-    box-sizing: border-box;
-`;
