+++ src/components/ui/hook-form/common.tsx
@@ -1,0 +1,13 @@
+import styled from '@emotion/styled';
+import { Typography } from 'antd';
+
+export const Label = styled(Typography.Text)`
+    margin-bottom: 10px;
+    font-size: 18px;
+    font-weight: 400;
+`;
+
+export const ErrorMessage = styled(Typography.Text)`
+    margintop: '5px';
+    color: red;
+`;
