+++ src/components/ui/hook-form/index.ts
@@ -2,4 +2,3 @@ export { FormDatePicker } from './datepicker';
 export { FormInput } from './input';
 export { FormTextarea } from './textarea';
 export { TimeSlots } from './timeslots';
-export { Button } from './button';
