+++ src/components/widgets/lang-switch/index.tsx
@@ -1,40 +1,34 @@
-import styled from '@emotion/styled';
-import Link from 'next/link';
+import { ConfigProvider, Segmented, SegmentedProps } from 'antd';
+import { usePathname } from 'next/navigation';
 import { useRouter } from 'next/router';
 import { COLORS } from 'src/styles/variables';
 
-const LangSwitchBase = ({ className }: { className?: string }) => {
-    const { locales, asPath } = useRouter();
+export const LangSwitch = () => {
+    const { locales, locale: currentLocale, push } = useRouter();
+    const currentPathname = usePathname();
+
+    const onLocaleChange: SegmentedProps['onChange'] = (value) => {
+        push(currentPathname, currentPathname, { locale: value as string });
+    };
 
     return (
-        <div className={className}>
-            {locales?.map((locale) => (
-                <Link
-                    className="link"
-                    key={locale}
-                    href={asPath}
-                    locale={locale}
-                >
-                    {locale}
-                </Link>
-            ))}
-        </div>
+        <ConfigProvider
+            theme={{
+                components: {
+                    Segmented: {
+                        trackBg: COLORS.blue,
+                        itemHoverBg: COLORS.blueHovered,
+                        itemColor: COLORS.colorInactive,
+                        itemHoverColor: COLORS.colorInactive,
+                    },
+                },
+            }}
+        >
+            <Segmented
+                options={locales as string[]}
+                value={currentLocale}
+                onChange={onLocaleChange}
+            />
+        </ConfigProvider>
     );
 };
-
-export const LangSwitch = styled(LangSwitchBase)`
-    .link {
-        padding: 4px 8px;
-        background: ${COLORS.white};
-        height: auto;
-        box-sizing: border-box;
-        border-radius: 5px;
-        color: ${COLORS.blue};
-        margin-right: 5px;
-        font-weight: 500;
-
-        &:hover {
-            color: ${COLORS.pink};
-        }
-    }
-`;
