+++ src/styles/variables.ts
@@ -4,5 +4,7 @@ export const COLORS = {
     pink: '#ee99a8',
     blue: '#5e63e8',
     blue2: '#4096ff',
+    blueHovered: '#5559d1',
+    colorInactive: '#151633',
     grey: '#d9d9d9',
 };
